# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MediTime API
  description: |
    API REST complète pour **MediTime**, application de gestion de médicaments et suivi des traitements médicaux.

    ## 🔐 Authentification
    L'API utilise **Supabase Auth** avec tokens JWT Bearer.  
    La plupart des endpoints nécessitent le header `Authorization: Bearer <token>`.

    ### Obtenir un token
    1. Authentifiez-vous via Supabase Auth (email/password, Google, etc.)
    2. Récupérez le token JWT depuis la session Supabase
    3. Ajoutez-le au header : `Authorization: Bearer <votre_token>`

    ## 🏗 Architecture
    | Composant | Technologie |
    |-----------|-------------|
    | Frontend | React 19 (Vite) + Bootstrap 5 + Supabase Client |
    | Backend | Flask (Python) + PostgreSQL + Raw SQL (psycopg2) |
    | Base de données | PostgreSQL avec Row Level Security (RLS) |
    | Infrastructure | Docker + CapRover + Render |

    ## 🚀 Guide de démarrage
    1. `GET /api/status` — Vérifier la connexion
    2. `GET /api/user/sync` — Synchroniser votre profil
    3. `GET /api/calendars` — Lister vos calendriers
    4. `POST /api/calendars` — Créer un calendrier

    ## 📦 Codes de réponse
    | Code | Signification |
    |------|---------------|
    | 200 | Succès |
    | 400 | Requête invalide (données manquantes) |
    | 401 | Non authentifié |
    | 403 | Accès refusé |
    | 404 | Ressource non trouvée |
    | 500 | Erreur serveur |

  version: 2.0.0
  contact:
    name: MediTime Support
...
  license:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT Supabase Auth.
        
        **Format du header :**
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIs...
        ```
  schemas:
    Error:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
          description: Message d'erreur lisible
          example: "Erreur lors de la récupération des données"
        code:
          type: string
          description: Code d'erreur machine
          example: "CALENDAR_FETCH_ERROR"
        i18n_key:
          type: string
          description: Clé de traduction i18n
          example: "api.calendar.fetch_error"
        error:
          type: string
          description: Détails techniques (dev uniquement)
        status_code:
          type: integer
          example: 500
    Success:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
          example: "Opération réussie"
        code:
          type: string
          example: "OPERATION_SUCCESS"
        i18n_key:
          type: string
        data:
          type: object
          additionalProperties: true
...
    Notification:
      type: object
      properties:
        notification_id:
          type: integer
        notification_type:
          type: string
          enum:
            - low_stock
            - calendar_invitation
            - calendar_invitation_accepted
            - calendar_invitation_rejected
            - calendar_shared_deleted_by_receiver
            - calendar_shared_deleted_by_owner
        read:
          type: boolean
        timestamp:
          type: string
          format: date-time
        calendar_id:
          type: integer
          nullable: true
        calendar_name:
          type: string
          nullable: true
        sender_name:
          type: string
          nullable: true
        sender_email:
          type: string
          nullable: true
        sender_photo_url:
          type: string
          nullable: true
        medication_name:
          type: string
          nullable: true
        medication_qty:
          type: integer
          nullable: true
        accepted:
          type: boolean
          nullable: true
        token:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
...
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Lister mes notifications
      operationId: getNotifications
      responses:
        '200':
          description: Notifications récupérées
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notifications:
                            type: array
                            items:
                              $ref: '#/components/schemas/Notification'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notifications/{notification_id}:
    patch:
      tags: [Notifications]
      summary: Marquer une notification comme lue
      operationId: markNotificationRead
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Notification marquée comme lue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '404':
          description: Notification non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notifications/register-token:
    post:
      tags: [Notifications]
      summary: Enregistrer un token FCM
      operationId: registerFCMToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: Token FCM du device
      responses:
        '200':
          description: Token enregistré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Token manquant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ═══════════════════════════════════════════════════════════════════════════
  # TOKENS (Partage public)
  # ═══════════════════════════════════════════════════════════════════════════
...